requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain_community

# TO RUN THE PROGRAM USE COMMAND GIVEN BELOW
# streamlit run app.py

# uncomment to use huggingface llms
# huggingface-hub

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2